refactor: 마이 갤러리 hook 나눔
